code/deployment/api/requirements.txt
fastapi
uvicorn
catboost
mlflow
pydantic
pandas
pyyaml